add stateful module and move relevant methods into stateful/stateless module folders
