+++ components/nimbus/src/lib.rs
@@ -6,7 +6,6 @@ mod defaults;
 mod enrollment;
 mod evaluator;
 mod json;
-mod matcher;
 mod sampling;
 mod strings;
 mod targeting;
@@ -19,28 +18,21 @@ pub use enrollment::{EnrolledFeature, EnrollmentStatus};
 pub use error::{NimbusError, Result};
 #[cfg(debug_assertions)]
 pub use evaluator::evaluate_enrollment;
-pub use matcher::AppContext;
 pub use schema::*;
 pub use targeting::NimbusTargetingHelper;
 
 cfg_if::cfg_if! {
     if #[cfg(feature = "stateful")] {
-        mod behavior;
-        mod client;
-        mod dbcache;
-        mod updating;
+        pub mod stateful;
 
-        pub mod nimbus_client;
-        pub mod persistence;
-
-        pub use crate::nimbus_client::*;
+        pub use stateful::nimbus_client::*;
+        pub use stateful::matcher::AppContext;
         pub use remote_settings::RemoteSettingsConfig;
     } else {
-        pub mod stateless {
-            pub mod cirrus_client;
-        }
+        pub mod stateless;
 
-        pub use crate::stateless::cirrus_client::*;
+        pub use stateless::cirrus_client::*;
+        pub use stateless::matcher::AppContext;
     }
 }
