+++ components/nimbus/src/stateful/client/fs_client.rs
@@ -6,8 +6,8 @@
 //! (eg, for testing against experiments which are not deployed anywhere) and
 //! for tests.
 
-use crate::client::SettingsClient;
 use crate::error::Result;
+use crate::stateful::client::SettingsClient;
 use crate::Experiment;
 use std::ffi::OsStr;
 use std::fs::File;
