+++ components/nimbus/src/stateful/mod.rs
@@ -1,0 +1,13 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+pub mod behavior;
+pub mod client;
+pub mod dbcache;
+pub mod enrollment;
+pub mod evaluator;
+pub mod matcher;
+pub mod nimbus_client;
+pub mod persistence;
+pub mod updating;
