+++ components/nimbus/src/tests/stateful/test_persistence.rs
@@ -2,7 +2,9 @@
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use crate::{enrollment::ExperimentEnrollment, error::Result, persistence::*, Experiment};
+use crate::{
+    enrollment::ExperimentEnrollment, error::Result, stateful::persistence::*, Experiment,
+};
 use rkv::StoreOptions;
 use serde_json::json;
 use std::fs;
