+++ components/nimbus/src/tests/stateful/test_updating.rs
@@ -6,8 +6,7 @@
 #![allow(unused_imports)]
 
 use crate::error::Result;
-use crate::persistence::Database;
-use crate::updating::*;
+use crate::stateful::{persistence::Database, updating::*};
 use crate::Experiment;
 
 // This test crashes lmdb for reasons that make no sense, so only run it
